Finish review changes
