+++ .gitignore
@@ -38,3 +38,4 @@
 !/app/assets/builds/.keep
 
 /.idea/*
+.run
