+++ README.md
@@ -12,10 +12,6 @@ Run locally: `bundle install` then `rails server`
 Tests: `rake test:run` and `rake test:system`
 
 TODO:
-- Swap out Clearance for default Rails 8 auth
-  - Remove rspec
-  - Remove clearance
-  - Make auth work well
 - List features needed for first public version
 
 Tile graphics: [FluffyStuff/riichi-mahjong-tiles](https://github.com/FluffyStuff/riichi-mahjong-tiles)
