+++ test/controllers/sessions_controller_test.rb
@@ -9,14 +9,16 @@ class SessionsControllerTest < ActionDispatch::IntegrationTest
 
   test "should create session with valid credentials" do
     assert_difference("Session.count", 1) do
-      post session_url, params: {email_address: users(:one).email_address, password: "password"}
+      post session_url, params: { email_address: users(:one).email_address, password: "password" }
     end
     assert_redirected_to root_url
     assert_signed_in
   end
 
   test "should not create session with invalid credentials" do
-    post session_url, params: {email_address: users(:one).email_address, password: "wrong_password"}
+    assert_no_difference("Session.count") do
+      post session_url, params: { email_address: users(:one).email_address, password: "wrong_password" }
+    end
     assert_redirected_to new_session_path
     assert_signed_out
     assert_not_nil flash[:alert]
@@ -24,11 +26,13 @@ class SessionsControllerTest < ActionDispatch::IntegrationTest
 
   test "should destroy session" do
     # First log in
-    post session_url, params: {email_address: users(:one).email_address, password: "password"}
+    post session_url, params: { email_address: users(:one).email_address, password: "password" }
     assert_signed_in
 
     # Then log out
-    delete session_url
+    assert_difference("Session.count", -1) do
+      delete session_url
+    end
     assert_redirected_to new_session_path
     assert_signed_out
   end
