+++ test/test_helper.rb
@@ -1,7 +1,6 @@
 ENV["RAILS_ENV"] ||= "test"
 require_relative "../config/environment"
 require "rails/test_help"
-require_relative "helpers/authentication_test_helper"
 require_relative "helpers/session_test_helper"
 
 module ActiveSupport
@@ -13,7 +12,6 @@ module ActiveSupport
     fixtures :all
 
     # Add more helper methods to be used by all tests here...
-    include AuthenticationTestHelper
     include SessionTestHelper
   end
 end
